[fix] Move helpers functions from reducers to a dedicated helpers library
